added navigation; added terms and conditions text; added forum/topic mark-as-read logic & tracking

+++ Serverless.Forum/Serverless.Forum/forum/PhpbbForumsTrack.cs
@@ -7,6 +7,6 @@ namespace Serverless.Forum.forum
     {
         public int UserId { get; set; }
         public int ForumId { get; set; }
-        public int MarkTime { get; set; }
+        public long MarkTime { get; set; }
     }
 }
